Search for missing blocks in chunks
The block catchup fetcher feels unresponsive because finding the missing
blocks over 6 millions block takes tens of minutes as `generate_series`
needs to generate an series with that many entries to diff against
the blocks table to find the missing block numbers.

To speed up the time until catch up starts fetching blocks, the search
range is reduced to 5,000 blocks at a time.  The SQL query has a number
restriction added to the blocks table, so the entire table isn't diffed
against a subset series to speed up the query.

The query is no longer allowed to take infinite time and instead uses
whatever the configured `timeout` is for the pool unless overridden.
This ensures that the query remain responsive.   We change the
`max_missing_block_count` instead of the
timeout if timeouts do occur.

Most of the important numbers have been moved out of the human-readable
log messages to metadata that is machine parsable to allow for
graphing it desired.